refacto: split dfa validator into own module
